google-cloud-related/cloudfunctions/trigger-github-worklfow/requirements.txt
google-api-python-client==2.146.0
requests==2.32.3
google-cloud-secret-manager==2.20.0
